add signature to topdown vote record
