+++ fendermint/vm/topdown/src/sync/mod.rs
@@ -138,7 +138,7 @@ where
 
     atomically(|| {
         view_provider.set_new_finality(finality.clone(), None)?;
-        vote_tally.set_finalized(finality.height, finality.block_hash.clone())?;
+        vote_tally.set_finalized(finality.height)?;
         vote_tally.set_power_table(power_table.clone())?;
         Ok(())
     })
